// src/app/models/puntos-mas-movil.enum.ts
export enum PuntosMasMovil {
  NUEVA_INTERIOR = 1.95,
  NUEVA_EXTERIOR = 2.18,
  NEBA_AUTOINST_INT = 1.95,
  NEBA_AUTOINST_EXT = 2.18,
  INSTALACION_POSTES = 4.13,
  REUTILIZADA = 1.33,
  REUTILIZADA_POSTES = 2.25,
  AVERIA_FESTIVO = 1.05,
  AVERIA = 1,
  AGILETV = 0.23,
  RECLAMACION = 0.97,
  NEBA = 1.1,
  GUARDIA_VALL_PAL = 0.0,
  REPETIDORES = 0.11,
  REUTILIZADA_ADAPTADORES = 1.38
}